test: migrate tests to integration-karma
