+++ packages/@lwc/engine/src/__tests__/traverse.spec.ts
@@ -7,342 +7,6 @@
 import { compileTemplate } from 'test-utils';
 import { createElement, LightningElement } from '../';
 
-describe('#childNodes', () => {
-    it('should always return an empty array for slots not rendering default content', () => {
-        const hasSlotTmpl = compileTemplate(`
-            <template>
-                <slot>
-                    <div></div>
-                </slot>
-            </template>
-        `);
-        class HasSlot extends LightningElement {
-            render() {
-                return hasSlotTmpl;
-            }
-        }
-
-        const parentTmpl = compileTemplate(
-            `
-            <template>
-                <div>
-                    <x-child-node-with-slot>
-                        <p></p>
-                        <p></p>
-                    </x-child-node-with-slot>
-                </div>
-            </template>
-        `,
-            {
-                modules: {
-                    'x-child-node-with-slot': HasSlot,
-                },
-            }
-        );
-        class Parent extends LightningElement {
-            render() {
-                return parentTmpl;
-            }
-        }
-
-        const elm = createElement('x-child-node-parent', { is: Parent });
-        document.body.appendChild(elm);
-        const slot = elm.shadowRoot
-            .querySelector('x-child-node-with-slot')
-            .shadowRoot.querySelector('slot');
-        // Per spec, slot elements will retain default content
-        expect(slot.childNodes).toHaveLength(1);
-        expect(slot.assignedNodes()).toHaveLength(2);
-    });
-
-    it('should return correct elements for slots rendering default content', () => {
-        const hasSlotTmpl = compileTemplate(`
-            <template>
-                <slot>
-                    <div></div>
-                </slot>
-            </template>
-        `);
-        class HasSlot extends LightningElement {
-            render() {
-                return hasSlotTmpl;
-            }
-        }
-
-        const parentTmpl = compileTemplate(
-            `
-            <template>
-                <div>
-                    <x-child-node-with-slot></x-child-node-with-slot>
-                </div>
-            </template>
-        `,
-            {
-                modules: {
-                    'x-child-node-with-slot': HasSlot,
-                },
-            }
-        );
-        class Parent extends LightningElement {
-            render() {
-                return parentTmpl;
-            }
-        }
-
-        const elm = createElement('x-child-node-parent', { is: Parent });
-        document.body.appendChild(elm);
-        const slot = elm.shadowRoot
-            .querySelector('x-child-node-with-slot')
-            .shadowRoot.querySelector('slot');
-        expect(slot.childNodes).toHaveLength(1);
-    });
-
-    it('should return correct elements for non-slot elements', () => {
-        const html = compileTemplate(`
-            <template>
-                <div>
-                    <p></p>
-                </div>
-            </template>
-        `);
-        class Parent extends LightningElement {
-            render() {
-                return html;
-            }
-        }
-
-        const elm = createElement('x-child-node-parent', { is: Parent });
-        document.body.appendChild(elm);
-        const child = elm.shadowRoot.querySelector('div');
-        const childNodes = child.childNodes;
-        expect(childNodes).toHaveLength(1);
-        expect(childNodes[0]).toBe(elm.shadowRoot.querySelector('p'));
-    });
-
-    it('should return correct elements for custom elements when no children present', () => {
-        const childTmpl = compileTemplate(`
-            <template>
-                <div></div>
-            </template>
-        `);
-        class Child extends LightningElement {
-            render() {
-                return childTmpl;
-            }
-        }
-
-        const parentTmpl = compileTemplate(
-            `
-            <template>
-                <div>
-                    <x-child></x-child>
-                </div>
-            </template>
-        `,
-            {
-                modules: {
-                    'x-child': Child,
-                },
-            }
-        );
-        class Parent extends LightningElement {
-            render() {
-                return parentTmpl;
-            }
-        }
-
-        const elm = createElement('x-child-node-parent', { is: Parent });
-        document.body.appendChild(elm);
-        const child = elm.shadowRoot.querySelector('x-child');
-        const childNodes = child.childNodes;
-        expect(childNodes).toHaveLength(0);
-    });
-
-    it('should return correct elements for custom elements when children present', () => {
-        const childTmpl = compileTemplate(`
-            <template>
-                <slot></slot>
-            </template>
-        `);
-        class Child extends LightningElement {
-            render() {
-                return childTmpl;
-            }
-        }
-
-        const parentTmpl = compileTemplate(
-            `
-            <template>
-                <div>
-                    <x-child>
-                        <p></p>
-                    </x-child>
-                </div>
-            </template>
-        `,
-            {
-                modules: {
-                    'x-child': Child,
-                },
-            }
-        );
-        class Parent extends LightningElement {
-            render() {
-                return parentTmpl;
-            }
-        }
-
-        const elm = createElement('x-child-node-parent', { is: Parent });
-        document.body.appendChild(elm);
-        const child = elm.shadowRoot.querySelector('x-child');
-        const childNodes = child.childNodes;
-        expect(childNodes).toHaveLength(1);
-    });
-
-    it('should return child text content passed via slot', () => {
-        const childTmpl = compileTemplate(`
-            <template>
-                <slot></slot>
-            </template>
-        `);
-        class Child extends LightningElement {
-            render() {
-                return childTmpl;
-            }
-        }
-
-        const parentTmpl = compileTemplate(
-            `
-            <template>
-                <div>
-                    <x-child>
-                        text
-                    </x-child>
-                </div>
-            </template>
-        `,
-            {
-                modules: {
-                    'x-child': Child,
-                },
-            }
-        );
-        class Parent extends LightningElement {
-            render() {
-                return parentTmpl;
-            }
-        }
-
-        const elm = createElement('x-child-node-parent', { is: Parent });
-        document.body.appendChild(elm);
-        const child = elm.shadowRoot.querySelector('x-child');
-        const childNodes = child.childNodes;
-        expect(childNodes).toHaveLength(1);
-        expect(childNodes[0].nodeType).toBe(3);
-        expect(childNodes[0].textContent).toBe('text');
-    });
-
-    it('should not return child text from within template', () => {
-        const childTmpl = compileTemplate(`
-            <template>
-                text
-            </template>
-        `);
-        class Child extends LightningElement {
-            render() {
-                return childTmpl;
-            }
-        }
-
-        const parentTmpl = compileTemplate(
-            `
-            <template>
-                <div>
-                    <x-child></x-child>
-                </div>
-            </template>
-        `,
-            {
-                modules: {
-                    'x-child': Child,
-                },
-            }
-        );
-        class Parent extends LightningElement {
-            render() {
-                return parentTmpl;
-            }
-        }
-
-        const elm = createElement('x-child-node-parent', { is: Parent });
-        document.body.appendChild(elm);
-        const child = elm.shadowRoot.querySelector('x-child');
-        const childNodes = child.childNodes;
-        expect(childNodes).toHaveLength(0);
-    });
-
-    it('should not return dynamic child text from within template', () => {
-        const childTmpl = compileTemplate(`
-            <template>
-                {dynamicText}
-            </template>
-        `);
-        class Child extends LightningElement {
-            get dynamicText() {
-                return 'text';
-            }
-            render() {
-                return childTmpl;
-            }
-        }
-
-        const parentTmpl = compileTemplate(
-            `
-            <template>
-                <x-child></x-child>
-            </template>
-        `,
-            {
-                modules: {
-                    'x-child': Child,
-                },
-            }
-        );
-        class Parent extends LightningElement {
-            render() {
-                return parentTmpl;
-            }
-        }
-
-        const elm = createElement('x-parent', { is: Parent });
-        document.body.appendChild(elm);
-        const childNodes = elm.shadowRoot.querySelector('x-child').childNodes;
-        expect(childNodes).toHaveLength(0);
-    });
-
-    it('should return correct childNodes from shadowRoot', () => {
-        const html = compileTemplate(`
-            <template>
-                <div></div>
-                text
-            </template>
-        `);
-        class Parent extends LightningElement {
-            render() {
-                return html;
-            }
-        }
-
-        const elm = createElement('x-child-node-parent', { is: Parent });
-        document.body.appendChild(elm);
-        const childNodes = elm.shadowRoot.childNodes;
-        expect(childNodes).toHaveLength(2);
-        expect(childNodes[0]).toBe(elm.shadowRoot.querySelector('div'));
-        expect(childNodes[1].nodeType).toBe(3);
-        expect(childNodes[1].textContent).toBe('text');
-    });
-});
-
 describe('assignedSlot', () => {
     it('should return null when custom element is not in slot', () => {
         class NoSlot extends LightningElement {}
