+++ packages/integration-karma/test/shadow-dom/Node-properties/x/dynamicText/dynamicText.html
@@ -1,0 +1,3 @@
+<template>
+    {dynamicText}
+</template>
