+++ packages/integration-karma/test/shadow-dom/Node-properties/x/dynamicText/dynamicText.js
@@ -1,0 +1,7 @@
+import { LightningElement } from 'lwc';
+
+export default class Child extends LightningElement {
+    get dynamicText() {
+        return 'text';
+    }
+}
