+++ packages/integration-karma/test/shadow-dom/Node-properties/x/hasNoSlot/hasNoSlot.html
@@ -1,0 +1,5 @@
+<template>
+    <div>
+        <p></p>
+    </div>
+</template>
