+++ packages/integration-karma/test/shadow-dom/Node-properties/x/hasNoSlot/hasNoSlot.js
@@ -1,0 +1,3 @@
+import { LightningElement } from 'lwc';
+
+export default class HasNoSlot extends LightningElement {}
