+++ packages/integration-karma/test/shadow-dom/Node-properties/x/hasSlot/hasSlot.html
@@ -1,0 +1,5 @@
+<template>
+    <slot>
+        <div></div>
+    </slot>
+</template>
