+++ packages/integration-karma/test/shadow-dom/Node-properties/x/hasSlot/hasSlot.js
@@ -1,0 +1,3 @@
+import { LightningElement } from 'lwc';
+
+export default class HasSlot extends LightningElement {}
