+++ packages/integration-karma/test/shadow-dom/Node-properties/x/parent/parent.html
@@ -1,0 +1,11 @@
+<template>
+    <div>
+        <x-has-no-slot></x-has-no-slot>
+    </div>
+    <div>
+        <x-text></x-text>
+    </div>
+    <div>
+        <x-dynamic-text></x-dynamic-text>
+    </div>
+</template>
