+++ packages/integration-karma/test/shadow-dom/Node-properties/x/parent/parent.js
@@ -1,0 +1,3 @@
+import { LightningElement } from 'lwc';
+
+export default class Parent extends LightningElement {}
