+++ packages/integration-karma/test/shadow-dom/Node-properties/x/simpleParent/simpleParent.html
@@ -1,0 +1,4 @@
+<template>
+    <div></div>
+    text
+</template>
