+++ packages/integration-karma/test/shadow-dom/Node-properties/x/slottedParent/slottedParent.html
@@ -1,0 +1,8 @@
+<template>
+    <div>
+        <x-has-slot>
+            <p></p>
+            <p></p>
+        </x-has-slot>
+    </div>
+</template>
