+++ packages/integration-karma/test/shadow-dom/Node-properties/x/textSlottedParent/textSlottedParent.html
@@ -1,0 +1,7 @@
+<template>
+    <div>
+        <x-has-slot>
+            text
+        </x-has-slot>
+    </div>
+</template>
