+++ packages/integration-karma/test/shadow-dom/Node-properties/x/unslottedParent/unslottedParent.html
@@ -1,0 +1,5 @@
+<template>
+    <div>
+        <x-has-slot></x-has-slot>
+    </div>
+</template>
